Fixed up routing, button and form display.
I was able to reformat the button and form to better display the images,
I just need to resize them and get a proper form graphic for the page to
be done. Couple possible ideas for a centered layered approach on the
form logic. Also I fixed up routing so all the pages now route to and
from each other properly.